requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
deprecation==2.1.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
google==3.0.0
google-auth==2.40.3
google-genai==1.21.1
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langsmith==0.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.5.0
networkx==3.5
numpy==2.3.1
orjson==3.10.18
packaging==24.2
pillow==11.2.1
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
realtime==2.4.3
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.3
supafunc==0.9.4
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
websockets==14.2
yarl==1.20.1
zstandard==0.23.0
